requirements.txt
aniso8601==8.0.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-RESTful==0.3.8
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.18.4
pandas==1.0.3
pyserial==3.4
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
six==1.14.0
urllib3==1.25.9
Werkzeug==1.0.1
